test: update mock stubs for IncrementFailedAttempts int return type
前コミットのインタフェース変更 (int64 → int) に合わせて、
各テストファイルのモック・スタブのシグネチャを更新する。

+++ backend/internal/usecase/admin/request_password_reset_test.go
@@ -39,7 +39,7 @@ func (m *mockAdminRepoForReqPwd) UpdatePassword(_ context.Context, _ int, _ stri
 }
 
 func (m *mockAdminRepoForReqPwd) Count(_ context.Context) (int, error) { return 0, nil }
-func (m *mockAdminRepoForReqPwd) IncrementFailedAttempts(_ context.Context, _ int) (int64, error) {
+func (m *mockAdminRepoForReqPwd) IncrementFailedAttempts(_ context.Context, _ int) (int, error) {
 	return 0, nil
 }
 func (m *mockAdminRepoForReqPwd) LockAccount(_ context.Context, _ int, _ time.Time) error {
